Move doesIdentityCookieExist into utils
